Adjust test to allow type variable as expression
This CL adjusts the test
language/constants_2018/constant_types_test
such that it expects no error when a type variable is used as a
potentially constant expression.

This feature was added for consistency as part of the constructor-
tearoffs feature bundle (because it can now be expressed anyway).

Change-Id: I5b58f2439699b24e3e3036cb0e3a0fcc9430d4c0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/214644
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Erik Ernst <eernst@google.com>