Deprecate ResultState.NOT_A_FILE, make the result VALID.
It duplicates `get exists`, and there are no other states.
All other states are presented by various InvalidResult(s).

We could have NotExistingFileResult and return it instead of a
ResolvedUnitResult, but ResolvedLibraryResult uses
List<ResolvedUnitResult> as `units`.

So, instead we say that it is valid, just does not exist,
and its content is empty.

Change-Id: I435555595590299df55d0016451a1e037934750d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/215900
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>