Prevent nulls from slipping into some error messages.
For CompileTimeErrorCode.UNDEFINED_METHOD, it was possible that the
undefined method reference might appear inside an unnamed extension
declaration.  This was fixed by changing the error message to refer to
the extended type rather than the extension name.

For HintCode.INVALID_VISIBLE_FOR_OVERRIDING_ANNOTATION and
CompileTimeErrorCode.EXTENSION_CONFLICTING_STATIC_AND_INSTANCE, we
just allow the name to be `<unnamed>` when the declaration in question
is an unnamed extension.

For AnalysisOptionsWarningCode.UNRECOGNIZED_ERROR_CODE and
AnalysisOptionsWarningCode.UNSUPPORTED_OPTION_WITH_LEGAL_VALUES, it's
actually ok for `null` to show up in the error message, because it can
only happen if the user has a literal `null` in their options file.
However, the way we want this to work is by passing the string `null`
as an argument to the error message (rather than the `null` value);
this will allow us to make message arguments non-nullable in a
follow-up CL.

Change-Id: I67138f314e6234d860c01d083d7c08a1a3ef11ed
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/216460
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>