Fix error messages that don't properly name members.
The error messages for INVALID_NON_VIRTUAL_ANNOTATION and
INVALID_SEALED_ANNOTATION appeared to reference the name of the
erroneously declared member, with messages like "The member '{0}'
can't be '@nonVirtual' because it isn't a concrete instance member.",
however the actual text that was substituted into the error message
wasn't the name of the member, but rather the name of the element
declaring the annotation.  So the error message would always be "The
member 'nonVirtual' can't be '@nonVirtual' because it isn't a concrete
instance member.", which was unhelpful and confusing.

Rather than fix the code to come up with the correct name of the
member (which would have been challenging, because the target of the
annotation could be an extension without a name, and wouldn't have
been that useful anyway, since the error location points to the
member), I've simply rephrased the error messages so that they don't
name the member in question.

Change-Id: I34f74e04607e16211b357429820c0336e0a03059
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/216563
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>