[gardening] Try to configure TSAN to produce coredumps on aborts
We seemingly don't get coredumps on the "iso-stress" builder when
running under TSAN. This is an attempt of setting TSAN settings via
an environment variable to explicitly force this behavior.

We also add a print to an existing test to help identify in which
section a crash happens.

TEST=Attempt to obtain cores on iso-stress when running under TSAN.

Change-Id: I698eb62e537923dc6733b4ed496a47f67338da34
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/217010
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>