[vm/concurrency] Remove vm/dart/regress_40753_test in iso-stress tests
This particular test causes thousands of exceptions to be thrown. This
is not a problem in itself, but it is when running under TSAN because it
will cause a buffer-overflow in TSAN's fixed-size shadow stack.

The buffer overflow will override data structures that often causes the
iso-stress runs to crash.

See more information in the issue.

Issue https://github.com/dart-lang/sdk/issues/47472#issuecomment-948235479

TEST=Should eliminate certain causes of crashes on "iso-stress" builder.

Change-Id: I9c3357eb49a31daf1cb728d745eeb405f2ff4a81
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/217600
Reviewed-by: Alexander Aprelev <aam@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>