Migration: fix crash when handling `await null`.
Previously, this wasn't working because the way we were checking
whether the type being awaited was a `Future` was via a subtype check
using the pre-NNBD type system.  And the pre-NNBD type system says
that `Null` is a subtype of `Future`.  So we were trying to extract a
type argument from the type `Null`, leading to a crash.

The solution is to add a separate case that directly handles the
`Null` type.

Bug: https://b.corp.google.com/issues/203199775
Change-Id: I083ef33e17c83d6a3e49b8fd0804c0de1c91ce84
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/218940
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>